XIVY-3255 move url download util back to its impl
