+++ src/main/java/ch/ivyteam/ivy/maven/InstallEngineMojo.java
@@ -213,7 +213,7 @@ public class InstallEngineMojo extends AbstractEngineMojo
     }
   }
 
-  private EngineDownloader getDownloader() throws MojoExecutionException
+  public EngineDownloader getDownloader() throws MojoExecutionException
   {
     if(downloadUsingMaven)
     {
