+++ src/main/java/ch/ivyteam/ivy/maven/engine/download/URLEngineDownloader.java
@@ -3,14 +3,21 @@ package ch.ivyteam.ivy.maven.engine.download;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
+import java.net.MalformedURLException;
 import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.StandardCopyOption;
+import java.util.Scanner;
+import java.util.regex.Pattern;
 
 import org.apache.commons.lang3.StringUtils;
+import org.apache.maven.artifact.versioning.ArtifactVersion;
+import org.apache.maven.artifact.versioning.DefaultArtifactVersion;
 import org.apache.maven.artifact.versioning.VersionRange;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.logging.Log;
+
+import ch.ivyteam.ivy.maven.engine.EngineVersionEvaluator;
 
 public class URLEngineDownloader implements EngineDownloader {
   private String zipFileName = null;
@@ -44,7 +51,7 @@ public class URLEngineDownloader implements EngineDownloader {
   {
     try (InputStream pageStream = new UrlRedirectionResolver().followRedirections(engineListPageUrl))
     {
-      return EngineDownloadURLResolver.findEngineDownloadUrl(pageStream, osArchitecture, ivyVersion, engineListPageUrl, ivyVersionRange);
+      return findEngineDownloadUrl(pageStream);
     }
     catch (IOException ex)
     {
@@ -99,6 +106,42 @@ public class URLEngineDownloader implements EngineDownloader {
       throw new IllegalStateException("Engine zip file name is not set up.");
     }
     return zipFileName;
+  }
+  
+  public URL findEngineDownloadUrl(InputStream htmlStream) throws MojoExecutionException, MalformedURLException
+  {
+      String engineFileNameRegex = "AxonIvyEngine[^.]+?\\.[^.]+?\\.+[^_]*?_"+osArchitecture+"\\.zip";
+      Pattern enginePattern = Pattern.compile("href=[\"|'][^\"']*?"+engineFileNameRegex+"[\"|']");
+      try(Scanner scanner = new Scanner(htmlStream))
+      {
+          String engineLink = null;
+          while (StringUtils.isBlank(engineLink))
+          {
+              String engineLinkMatch = scanner.findWithinHorizon(enginePattern, 0);
+              if (engineLinkMatch == null)
+              {
+                  throw new MojoExecutionException("Could not find a link to engine for version '"+ivyVersion+"' on site '"+engineListPageUrl+"'");
+              }
+              String versionString = StringUtils.substringBetween(engineLinkMatch, "AxonIvyEngine", "_"+osArchitecture);
+              ArtifactVersion version = new DefaultArtifactVersion(EngineVersionEvaluator.toReleaseVersion(versionString));
+              if (ivyVersionRange.containsVersion(version))
+              {
+                  engineLink = StringUtils.replace(engineLinkMatch, "\"", "'");
+                  engineLink = StringUtils.substringBetween(engineLink, "href='", "'");
+              }
+          }
+          return toAbsoluteLink(engineListPageUrl, engineLink);
+      }
+  }
+
+  private static URL toAbsoluteLink(URL baseUrl, String parsedEngineArchivLink) throws MalformedURLException
+  {
+      boolean isAbsoluteLink = StringUtils.startsWithAny(parsedEngineArchivLink, "http://", "https://");
+      if (isAbsoluteLink)
+      {
+          return new URL(parsedEngineArchivLink);
+      }
+      return new URL(baseUrl, parsedEngineArchivLink);
   }
 
 }
