+++ src/test/java/ch/ivyteam/ivy/maven/TestInstallEngineMojo.java
@@ -36,8 +36,8 @@ import org.junit.Rule;
 import org.junit.Test;
 
 import ch.ivyteam.ivy.maven.engine.EngineClassLoaderFactory.OsgiDir;
-import ch.ivyteam.ivy.maven.engine.download.EngineDownloadURLResolver;
 import ch.ivyteam.ivy.maven.engine.EngineVersionEvaluator;
+import ch.ivyteam.ivy.maven.engine.download.URLEngineDownloader;
 import mockit.Mock;
 import mockit.MockUp;
 import net.lingala.zip4j.core.ZipFile;
@@ -387,7 +387,7 @@ public class TestInstallEngineMojo
 
   private String findLink(String html) throws MojoExecutionException, MalformedURLException
   {
-    return EngineDownloadURLResolver.findEngineDownloadUrl(IOUtils.toInputStream(html), mojo.osArchitecture, mojo.ivyVersion, mojo.engineListPageUrl, mojo.getIvyVersionRange()).toExternalForm();
+    return getUrlDownloader().findEngineDownloadUrl(IOUtils.toInputStream(html)).toExternalForm();
   }
 
   @Test
@@ -396,11 +396,16 @@ public class TestInstallEngineMojo
     boolean run = Boolean.parseBoolean(System.getProperty("run.public.download.test"));
     assumeTrue("SKIPPING test 'testDefaultListPage_isAvailable'", run);
 
-    String engineUrl = EngineDownloadURLResolver.findEngineDownloadUrl(mojo.engineListPageUrl.openStream(), mojo.osArchitecture, mojo.ivyVersion, mojo.engineListPageUrl, mojo.getIvyVersionRange()).toExternalForm();
+    String engineUrl = getUrlDownloader().findEngineDownloadUrl(mojo.engineListPageUrl.openStream()).toExternalForm();
     assertThat(engineUrl)
       .as("The default engine list page url '"+mojo.engineListPageUrl.toExternalForm()+"' "
               + "must provide an engine for the current default engine version '"+mojo.ivyVersion+"'.")
       .contains(mojo.ivyVersion);
+  }
+
+  private URLEngineDownloader getUrlDownloader() throws MojoExecutionException
+  {
+    return (URLEngineDownloader)mojo.getDownloader();
   }
 
   @Test
